services/email-service/python-flask-server/requirements.txt
connexion >= 2.6.0
connexion[swagger-ui] >= 2.6.0
python_dateutil == 2.6.0
setuptools >= 21.0.0
swagger-ui-bundle >= 0.0.2
flask_cors >= 3.0.10
sendgrid >= 6.9.7